requirements.txt
smolagents>=0.4.0
openai>=1.0.0
anthropic>=0.25.0
google-generativeai>=0.5.0
langchain>=0.1.0
langchain-community>=0.0.20
langchain-huggingface>=0.0.3
langchain-openai>=0.1.0
langchain-anthropic>=0.1.0
langchain-google-genai>=1.0.0
faiss-cpu>=1.7.4
PyPDF2>=3.0.0
pymupdf>=1.23.0
sentence-transformers>=2.2.0
duckduckgo-search>=5.0.0
gradio>=4.0.0
gradio-tools>=0.0.9
python-dotenv>=1.0.0
tqdm>=4.65.0
markdownify>=0.11.6
spaces>=0.18.0
datasets>=2.14.0
transformers>=4.35.0
torch>=2.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
